Move technology-items.service to services/ dir
